# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateDepartmentRequest:
    properties:
      code:
        maxLength: 50
        type: string
      description:
        maxLength: 255
        type: string
      leader_id:
        type: string
      name:
        maxLength: 100
        type: string
      parent_id:
        type: string
      sort_order:
        type: integer
    required:
    - code
    - name
    type: object
...
  dto.CreateTemplateRequest:
    properties:
      body_template:
        type: string
      category:
        type: string
      channels:
        items:
          type: string
        type: array
      code:
        type: string
      name:
        type: string
      title_template:
        type: string
      variables_schema:
        additionalProperties:
          type: string
        type: object
    required:
    - body_template
    - channels
    - code
    - name
    - title_template
    type: object
...
  dto.DepartmentResponse:
    properties:
      children:
        items:
          $ref: '#/definitions/dto.DepartmentTreeResponse'
        type: array
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      leader_id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      sort_order:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
  dto.DepartmentTreeResponse:
    properties:
      children:
        items:
          $ref: '#/definitions/dto.DepartmentTreeResponse'
        type: array
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      leader_id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      sort_order:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.TestTemplateRequest:
    properties:
      variables:
        additionalProperties: true
        type: object
    type: object
...
  dto.UpdateDepartmentRequest:
    properties:
      description:
        maxLength: 255
        type: string
      leader_id:
        type: string
      name:
        maxLength: 100
        type: string
      sort_order:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    type: object
...
  dto.UpdateTemplateRequest:
    properties:
      body_template:
        type: string
      category:
        type: string
      channels:
        items:
          type: string
        type: array
      name:
        type: string
      status:
        type: integer
      title_template:
        type: string
      variables_schema:
        additionalProperties:
          type: string
        type: object
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      request_id:
        type: string
      timestamp:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: 高校计算机协会一体化管理平台
  title: StarByte API
  version: "1.0"
paths:
...
  /system/departments:
    get:
      description: 获取完整的部门树形结构
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.DepartmentTreeResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取部门树
      tags:
      - 部门管理
    post:
      consumes:
      - application/json
      description: 创建新部门
      parameters:
      - description: 部门信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DepartmentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建部门
      tags:
      - 部门管理
  /system/departments/{id}:
    delete:
      description: 删除指定部门
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除部门
      tags:
      - 部门管理
    get:
      description: 根据ID获取部门详细信息
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DepartmentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取部门详情
      tags:
      - 部门管理
    put:
      consumes:
      - application/json
      description: 更新部门信息
      parameters:
      - description: 部门ID
        in: path
        name: id
        required: true
        type: string
      - description: 部门信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DepartmentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新部门
      tags:
      - 部门管理
...
  /system/notification-templates:
    get:
      description: 获取通知模板列表（分页）
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 模板列表
      tags:
      - 系统管理-通知模板
    post:
      consumes:
      - application/json
      description: 创建通知模板（管理员）
      parameters:
      - description: 模板信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建通知模板
      tags:
      - 系统管理-通知模板
  /system/notification-templates/{id}:
    delete:
      description: 删除通知模板
      parameters:
      - description: 模板 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除模板
      tags:
      - 系统管理-通知模板
    get:
      description: 获取通知模板详情
      parameters:
      - description: 模板 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 模板详情
      tags:
      - 系统管理-通知模板
    put:
      consumes:
      - application/json
      description: 更新通知模板
      parameters:
      - description: 模板 ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新模板
      tags:
      - 系统管理-通知模板
  /system/notification-templates/{id}/test:
    post:
      consumes:
      - application/json
      description: 使用提供的变量测试渲染通知模板
      parameters:
      - description: 模板 ID
        in: path
        name: id
        required: true
        type: string
      - description: 测试变量
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TestTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 测试模板
      tags:
      - 系统管理-通知模板
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"